+++ src/main/java/de/jeff_media/Drop2InventoryPlus/listeners/GenericListener.java
@@ -11,6 +11,7 @@ import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
 import org.bukkit.event.block.BlockBreakEvent;
 import org.bukkit.event.entity.EntityDamageByEntityEvent;
+import org.bukkit.event.entity.EntityDamageEvent;
 import org.bukkit.event.entity.EntityDeathEvent;
 import org.bukkit.event.hanging.HangingBreakByEntityEvent;
 import org.bukkit.event.player.PlayerJoinEvent;
@@ -90,6 +91,22 @@ public class GenericListener implements org.bukkit.event.Listener {
         if (!main.utils.isMobEnabled(victim)) {
             main.debug("not enabled for victim type "+victim.getType().name());
             return;
+        }
+
+        if(main.getConfig().getBoolean(Config.IGNORE_DROPS_FROM_MOBS_KILLED_BY_LAVA)) {
+            if(event.getEntity().getLastDamageCause().getCause()== EntityDamageEvent.DamageCause.LAVA) {
+                main.debug("ignore drops from mobs killed by lava: true");
+                main.legacyDropDetectionManager.registerIgnoredLocation(event.getEntity().getLocation());
+                return;
+            }
+        }
+
+        if(main.getConfig().getBoolean(Config.IGNORE_DROPS_FROM_MOBS_KILLED_BY_LAVA)) {
+            if(event.getEntity().getLastDamageCause().getCause()== EntityDamageEvent.DamageCause.HOT_FLOOR) {
+                main.debug("ignore drops from mobs killed by magma: true");
+                main.legacyDropDetectionManager.registerIgnoredLocation(event.getEntity().getLocation());
+                return;
+            }
         }