+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.2.0
+- Added new config options:
+  - ignore-drops-from-mobs-killed-by-lava: false
+  - ignore-drops-from-mobs-killed-by-magma: false
+
 ## 1.1.2
 - Improved automatic update checker